# Add bicubic resize for RGB images