… tests/test-resize1.png
… tests/test-resize2.png
… tests/test-resize3.png
… tests/test-resize4.png