This closes #9 by adding cubic interpolation option to `resize` (currently only implemented for RGB images - TODO for everything else).

This implementation is slow and conses like crazy:

```lisp
CL-USER> (let* ((image (imago:read-png "~/Downloads/hatsu2.png")))
           (time (imago:resize image 900 1200 :interpolation 'imago:bicubic)))
Evaluation took:
  6.099 seconds of real time
  6.108628 seconds of total run time (6.106142 user, 0.002486 system)
  [ Run times consist of 0.102 seconds GC time, and 6.007 seconds non-GC time. ]
  100.16% CPU
  21,356,979,167 processor cycles
  10,234,163,152 bytes consed
  
#<IMAGO:RGB-IMAGE (900x1200) {10192994F3}>
```

But I want to get the algorithm correct and pin it with some test cases before I start optimizing it away. Are there any test suites for image resizing with bicubic interpolation that you are aware of?

------------

Showcase:

Original:
![hatsu2](https://user-images.githubusercontent.com/15045546/147513861-d6bb620a-a948-4e3b-b4fc-f88a1f1cd3a0.png)

`(imago:resize image 150 200 :interpolation 'imago:bicubic)` (shrunk):
![test11](https://user-images.githubusercontent.com/15045546/147513877-708dcc9d-5175-4c6e-8b33-2d7aa4ed73cd.png)

`(imago:resize image 900 1200 :interpolation 'imago:bicubic)` (enlarged):
![test10](https://user-images.githubusercontent.com/15045546/147513919-5e77778a-efa2-450c-8cb8-4623209aa403.png)