@shamazmazum 2021-12-28
I did what I planned.

Can you rewrite your contribution so it adds a new `interpolate-pixel` method which takes an image and coordinates `(x, y)` in `(single-float 0f0 1f0)` format, where zero means upper left and one means lower right corners of the image?

When you do so, you will be able to call `(resize image new-width new-height :interpolation :your-new-interpolation-method)`.