@phoe 2021-12-28
OK, thanks - I'll spend a short while on making the algorithm a little bit more acceptable time-wise and I'll rebase to your current master.

Should interpolation methods be named after keywords or external symbols in the `imago` package?