@shamazmazum 2021-12-28
> Should interpolation methods be named after keywords or external symbols in the imago package?

IMHO, symbols from `keword` package are better suited for this purpose.