@phoe 2021-12-28
OK, the resize is now 20 times not as damn slow as before. Let me update to the new API of yours.