@phoe 2021-12-28
Fully rebased and ready for review.

Performance is still slow but a little bit better; I think that any future improvement will require indulging in some vector operations unless I badly missed something in the algorithm.

```lisp
CL-USER> (let* ((image (imago:read-png "~/Downloads/hatsu2.png")))
           (time (imago:resize image 900 1200 :interpolation :bicubic)))
Evaluation took:
  0.319 seconds of real time
  0.321307 seconds of total run time (0.321282 user, 0.000025 system)
  100.63% CPU
  1,124,300,583 processor cycles
  4,320,016 bytes consed
  
#<IMAGO:RGB-IMAGE (900x1200) {1011370C33}>
```