@phoe 2021-12-28
TODO: use `alex:array-index` as a type for array indices.