@shamazmazum 2021-12-28
Thanks! Few things to fix before merging.

1) Can you please use `git reset` and/or `git rebase` to produce one or two commits which are easier to review? You can force-push then.
2) I renamed `:linear` to `:bilinear` as you asked, this PR seems to rename it back.
3) Please do not operate with pixels as if they were mere integers (even if they really are). Do not call `ash`, `ldb`, `logfoo` on them. Use pixel constructors `make-gray`, `make-color` and accessors `color-red`, `color-green`, etc. and `gray-intensity`, `gray-alpha`. These functions are there for a reason, after all :)