@phoe 2021-12-28
1. Done - squashed all commits into one.
2. Thanks for the catch! Fixed.
3. OK, I'll try to restructure my code.