@phoe 2021-12-28
Change #3 results in a 60% slowdown (~300ms vs ~500ms on my test file). Are you sure you want to require me to stick to accessor functions?

Old version (byte munging):

```lisp
CL-USER> (let* ((image (imago:read-png "~/Downloads/hatsu2.png"))) 
           (the-cost-of-nothing:bench
            (imago:resize image 900 1200 :interpolation :bicubic)))
294.10 milliseconds
; No value
```

New version (accessors):

```lisp
CL-USER> (let* ((image (imago:read-png "~/Downloads/hatsu2.png"))) 
           (the-cost-of-nothing:bench
            (imago:resize image 900 1200 :interpolation :bicubic)))
489.99 milliseconds
; No value
```

The changed Lisp code:

```lisp
(defun interpolate-pixel-cubic (image x y constructor accessors)
  (declare (optimize speed (debug 0)))
  (declare (type (single-float 0f0 1f0) x y))
  (declare (type function constructor))
  (with-image-definition (image width height image-pixels)
    (declare (type (unsigned-byte 32) height width))
    (declare (type (simple-array (unsigned-byte 32) (* *)) image-pixels))
    (let ((x (- (* x width) 0.5f0))
          (y (- (* y height) 0.5f0)))
      (declare (type (single-float -10f10 10f10) x y))
      (labels ((compute-neighborhood (real-y real-x neighborhood)
                 (declare (inline clamp))
                 (dotimes (i (length +resize-offsets+))
                   (let* ((cell (aref +resize-offsets+ i))
                          (offset-y (car cell))
                          (offset-x (cdr cell)))
                     (declare (type (signed-byte 8) offset-x offset-y))
                     (let ((y (ldb (byte 32 0) (+ real-y offset-y)))
                           (x (ldb (byte 32 0) (+ real-x offset-x))))
                       (setf (aref neighborhood i)
                             (aref image-pixels
                                   (clamp y 0 (1- height))
                                   (clamp x 0 (1- width))))))))
               (cubic (x p0 p1 p2 p3)
                 (declare (type single-float x p0 p1 p2 p3))
                 (let* ((step1 (* x (+ (* 3f0 (- p1 p2)) p3 (- p0))))
                        (step2 (* x (+ (* 2f0 p0) (* -5f0 p1)
                                       (* 4f0 p2) (- p3) step1)))
                        (step3 (* x (+ p2 (- p0) step2))))
                   (+ p1 (* 0.5f0 step3))))
               (bicubic (x y pixels accessor)
                 (declare (type single-float x y))
                 (declare (type (simple-array (unsigned-byte 32) (16)) pixels))
                 (declare (type function accessor))
                 (macrolet ((px (n) `(float (the (unsigned-byte 32)
                                                 (funcall accessor (aref pixels ,n)))
                                            1f0)))
                   (let ((x1 (cubic y (px 0) (px 1) (px 2) (px 3)))
                         (x2 (cubic y (px 4) (px 5) (px 6) (px 7)))
                         (x3 (cubic y (px 8) (px 9) (px 10) (px 11)))
                         (x4 (cubic y (px 12) (px 13) (px 14) (px 15))))
                     (the (single-float -10f10 10f10) (cubic x x1 x2 x3 x4))))))
        (declare (inline compute-neighborhood cubic bicubic))
        (multiple-value-bind (real-x interp-x) (truncate x)
          (multiple-value-bind (real-y interp-y) (truncate y)
            (let ((neighborhood (make-array 16 :element-type '(unsigned-byte 32))))
              (declare (dynamic-extent neighborhood))
              (compute-neighborhood real-y real-x neighborhood)
              (flet ((access (accessor)
                       (let* ((interpolated (bicubic interp-y interp-x neighborhood accessor))
                              (clamped (clamp (round interpolated) 0 255)))
                         clamped)))
                (apply constructor (mapcar #'access accessors)))
              ;; (let ((result '()))
              ;;   (dolist (accessor accessors (apply constructor result))
              ;;     (declare (type (unsigned-byte 8) bytepos))
              ;;     (let* ((interpolated (bicubic interp-y interp-x neighborhood))
              ;;            (accessed (funcall accessor interpolated))
              ;;            (clamped (clamp (round accessed) 0 255)))
              ;;       (declare (type (unsigned-byte 8) clamped)
              ;;                (type (integer 0 32) bytepos))
              ;;       (incf result shifted))))
              )))))))

(defmethod interpolate-pixel ((image rgb-image) x y (method (eql :bicubic)))
  (interpolate-pixel-cubic image x y #'make-color
                           (list #'color-red #'color-green
                                 #'color-blue #'color-alpha)))

(defmethod interpolate-pixel ((image grayscale-image) x y (method (eql :bicubic)))
  (interpolate-pixel-cubic image x y #'make-gray
                           (list #'gray-intensity #'gray-alpha)))
```