@shamazmazum 2021-12-29
Do not worry about slowdown. I'll optimize what can be optimized after merging. A clean code is a priority